% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fine_tune.R
\name{fine_tune}
\alias{fine_tune}
\title{Return an upated communtiy assignments and
memberships by k-means clustering.}
\usage{
fine_tune(label, r)
}
\arguments{
\item{label}{Current labels of community assignments.}

\item{r}{A p-dimensional vector of eigenvector times
square root of non-negative eigenvalues, where p equals
the number of non-negative eigenvalues.}
}
\value{
Returns a data frame with entries:
\describe{
  \item{mem_up:}{ An updated measure of the strenth with
  which vertices assigned to their new communities. }
  \item{mem_norm_up:}{ a softMax normalized "mem_up".}
  \item{label_up:}{ Updated labels of community assignments.}
}
}
\description{
Return an upated communtiy assignments and
memberships by k-means clustering.
}
\examples{
\donttest{
fine_tune(label, r)
}
}
